requirements.txt
sqlalchemy==1.4.44
pydantic==1.10.2
email-validator==1.3.0
psycopg2-binary==2.9.5
fastapi==0.87.0
alembic==1.8.1
uvicorn==0.20.0
bcrypt==4.0.1
passlib==1.7.4
pytz
python-multipart==0.0.5
rstr==3.2.0
python-jose==3.3.0

